update card grid connect to data
