+++ client/src/component/card/CardFrame.js
@@ -1,0 +1,68 @@
+import React, { useEffect } from 'react';
+import AppBar from '@material-ui/core/AppBar';
+import Button from '@material-ui/core/Button';
+import LibraryBooksIcon from '@material-ui/icons/LibraryBooks';
+import Card from '@material-ui/core/Card';
+import CardActions from '@material-ui/core/CardActions';
+import CardContent from '@material-ui/core/CardContent';
+import CardMedia from '@material-ui/core/CardMedia';
+import CssBaseline from '@material-ui/core/CssBaseline';
+import Grid from '@material-ui/core/Grid';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
+import { makeStyles } from '@material-ui/core/styles';
+import Container from '@material-ui/core/Container';
+import Link from '@material-ui/core/Link';
+import {NAVER_COLOR} from '../../models/colors';
+
+function CardFrame() {
+  const classes = useStyles();
+
+  return (
+    <Card className={classes.card}>
+      <CardMedia
+        className={classes.cardMedia}
+        image="https://source.unsplash.com/random"
+        title="Image title"
+      />
+      <CardContent className={classes.cardContent}>
+        <Typography gutterBottom variant="h5" component="h2">
+          Heading
+        </Typography>
+        <Typography>
+          This is a media card. You can use this section to describe the content.
+        </Typography>
+      </CardContent>
+      <CardActions>
+        <Button size="small" color="primary">
+          View
+        </Button>
+        <Button size="small" color="primary">
+          Edit
+        </Button>
+      </CardActions>
+    </Card>
+  );
+}
+
+const useStyles = makeStyles((theme) => ({
+  icon: {
+    marginRight: theme.spacing(2),
+  },
+  cardGrid: {
+    paddingTop: theme.spacing(8),
+    paddingBottom: theme.spacing(8),
+  },
+  card: {
+    height: '100%',
+    display: 'flex',
+    flexDirection: 'column',
+  },
+  cardMedia: {
+    paddingTop: '56.25%', // 16:9
+  },
+  cardContent: {
+    flexGrow: 1,
+  },
+}));
+export default CardFrame;
